Some cleanup and code review of compr package.
